Fix unwanted PID detection in input_docsis. Don't error out in target_pcap if layer wasn't found in packet but simply issue a warning and continue. Fix small memory leak in xmlrpccmd_list_loaded_target(). Make sure we use ssize_t and not size_t for read()s. Disable nasty XMLRPC_IPV6 hack. Fix multiple bugs in target_msn_session_merge(). Make sure conntrack privs have a conversation created when receiving SIP messages in target_msn.

+++ src/mgmtsrv.c
@@ -267,7 +267,7 @@ int mgmtsrv_accept_connection(struct mgmt_connection *c) {
 int mgmtsrv_read_socket(struct mgmt_connection *c) {
 
 	
-	int res = 0;
+	ssize_t res = 0;
 	unsigned int len = 0;
 	unsigned char buffer[READ_BUFF_LEN];
